BREAKING CHANGE: change file structure for routing, separation of layout folder from root to auth
