+++ app/(main)/layout.tsx
@@ -1,0 +1,30 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "../globals.css";
+import Navbar from "../ui/navbar";
+import Footer from "../ui/footer";
+import CheckAvailability from "../ui/check-availability";
+import { outfit } from "../ui/fonts";
+
+const inter = Inter({ subsets: ["latin"], variable: "--font-sans" });
+
+export const metadata: Metadata = {
+  title: "Zenglobal",
+  description: "Find deals on fast, low-cost and reliable internet near you",
+};
+
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body className={`${outfit.variable} font-sans`}>
+        <Navbar />
+        {children}
+        <Footer />
+      </body>
+    </html>
+  );
+}
