+++ app/(main)/our-mission/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Hero from "../ui/our-mission/hero";
-import ShareTheWord from "../ui/our-mission/share-the-word";
+import Hero from "../../ui/our-mission/hero";
+import ShareTheWord from "../../ui/our-mission/share-the-word";
 
 export default function Page() {
   return (
