+++ app/(main)/policies/customer-agreements-and-policies/page.tsx
