+++ app/(main)/policies/social-responsibility/page.tsx
