--- app/auth/signin/page.tsx
@@ -1,8 +1,0 @@
-import SignInForm from '@/app/ui/auth/signin/form'
-import React from 'react'
-
-export default function page() {
-  return (
-    <SignInForm />
-  )
-}
