+++ app/layout.tsx
@@ -1,12 +1,5 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
-import Navbar from "./ui/navbar";
-import Footer from "./ui/footer";
-import CheckAvailability from "./ui/check-availability";
-import { outfit } from "./ui/fonts";
-
-const inter = Inter({ subsets: ["latin"], variable: "--font-sans" });
 
 export const metadata: Metadata = {
   title: "Zenglobal",
@@ -20,13 +13,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={`${outfit.variable} font-sans`}>
-        <Navbar />
-        {/* <CheckAvailability /> */}
-
-        {children}
-        <Footer />
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
